+++ database/migrations/2022_03_21_000000_create_ip_addresses_table.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('ip_addresses', function (Blueprint $table) {
+            $table->id();
+            $table->foreignBigInteger('user_id', 'users');
+            $table->integer('increment')->default(1);
+            $table->string('credential');
+            $table->binary('ip_address');
+            $table->boolean('is_verified');
+            $table->timestamp('created_at')->nullable();
+            $table->timestamp('updated_at')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('ip_addresses');
+    }
+};